requirements.txt
whitenoise==6.4.0
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-filter==23.2
django-phonenumber-field==7.1.0
django-rename-app==0.1.6
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-writable-nested==0.7.0
filelock==3.12.0
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
packaging==23.1
phonenumbers==8.13.11
Pillow==9.5.0
platformdirs==3.5.0
pluggy==1.0.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.6.0
pyproject_api==1.5.1
python3-openid==3.2.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tox==4.5.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.23.0
django-environ==0.10.0
